remove unnecessary logic
