+++ novawallet/Modules/Staking/Services/RewardCalculatorService/RelayChain/InflationCurveRewardEngine.swift
@@ -23,11 +23,22 @@ final class InflationCurveRewardEngine: RewardCalculatorEngine {
         }
     }
 
+    // We are solving equation to find era interest - x: yr * T = T * (1 + x)^t - T
     override func calculateEraReturn(from annualReturn: Decimal) -> Decimal {
-        solveExponential(
-            for: annualReturn,
-            eraDurationInSeconds: eraDurationInSeconds,
-            daysInYear: TimeInterval(CalculationPeriod.year.inDays)
-        )
+        guard eraDurationInSeconds > 0 else {
+            return 0
+        }
+
+        let daysInYear = TimeInterval(CalculationPeriod.year.inDays)
+        let erasInYear = daysInYear * TimeInterval.secondsInDay / eraDurationInSeconds
+
+        guard erasInYear > 0 else {
+            return 0
+        }
+
+        let rawAnnualReturn = (annualReturn as NSDecimalNumber).doubleValue
+        let result = pow(rawAnnualReturn + 1.0, 1.0 / Double(erasInYear)) - 1.0
+
+        return Decimal(result)
     }
 }
